requirements.txt
(Python==3.7.6)
biosppy>=0.6.1
Click>=7.0
heartpy>=1.2.6
matplotlib>=3.1.3
mne>=0.20.5
numpy>=1.18.1
pandas>=1.0.1
plotly>=4.5.0
py-ecg-detectors>=1.0.2
streamlit>=0.56.0
wfdb>=2.2.1
